+++ src/stactools/core/__init__.py
@@ -20,4 +20,4 @@ __all__ = [
     "move_all_assets",
     "use_fsspec",
 ]
-__version__ = "0.2.4"
+__version__ = "0.2.5"